+++ src/r5py/r5/__init__.py
@@ -3,6 +3,7 @@
 """R5 classes."""
 
 from .access_leg import AccessLeg
+from .breakdown_stat import BreakdownStat
 from .detailed_itineraries import DetailedItineraries
 from .direct_leg import DirectLeg
 from .egress_leg import EgressLeg
@@ -16,11 +17,13 @@ from .transit_leg import TransitLeg
 from .transport_mode import TransportMode
 from .transport_network import TransportNetwork
 from .travel_time_matrix import TravelTimeMatrix
+from .travel_time_matrix_decomposed import TravelTimeMatrixDecomposed
 from .trip import Trip
 from .trip_planner import TripPlanner
 
 __all__ = [
     "AccessLeg",
+    "BreakdownStat",
     "DetailedItineraries",
     "DirectLeg",
     "EgressLeg",
@@ -34,6 +37,7 @@ __all__ = [
     "TransportMode",
     "TransportNetwork",
     "TravelTimeMatrix",
+    "TravelTimeMatrixDecomposed",
     "Trip",
     "TripPlanner",
 ]